apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
mangum==0.19.0
stripe==11.3.0
python-multipart==0.0.19
